Add 'New article'-button; change default number of visible elements

+++ src/components/tina/ArticleTable.tsx
@@ -1,22 +1,25 @@
-import { ArticleTableData } from "../../utils/ArticleDB";
-import { useRouter } from "next/router";
-import MaterialTable, { Action } from "material-table";
-import { ArticleAPI } from "../../utils/ArticleAPI";
-import { useState } from "react";
+import { ArticleTableData } from "../../utils/ArticleDB"
+import { useRouter } from "next/router"
+import MaterialTable from "material-table"
+import { ArticleAPI } from "../../utils/ArticleAPI"
+import { useState } from "react"
+import { useSession } from "next-auth/client"
 
 export interface ArticleTableProps {
-  articles: ArticleTableData[];
+  articles: ArticleTableData[]
 }
 
 export const ArticleTable = ({ articles }: ArticleTableProps): JSX.Element => {
-  const [tableData, setTableData] = useState(articles);
+  const [session] = useSession()
+  const [tableData, setTableData] = useState(articles)
 
-  const router = useRouter();
+  const router = useRouter()
 
-  const api = new ArticleAPI();
+  const api = new ArticleAPI()
 
   return (
     <MaterialTable
+      title={`Hello ${session ? session.user.name : ""}`}
       columns={[
         { title: "ID", field: "id" },
         { title: "Title", field: "title" },
@@ -32,8 +35,8 @@ export const ArticleTable = ({ articles }: ArticleTableProps): JSX.Element => {
           icon: "edit",
           tooltip: "Edit article",
           onClick: (_, rowData) => {
-            const data = rowData as ArticleTableData;
-            router.push(`/tina/${data.url}`);
+            const data = rowData as ArticleTableData
+            router.push(`/tina/${data.url}`)
           },
         },
         {
@@ -41,23 +44,34 @@ export const ArticleTable = ({ articles }: ArticleTableProps): JSX.Element => {
           icon: "delete",
           tooltip: "Delete article",
           onClick: async (_, rowData) => {
-            const data = rowData as ArticleTableData;
+            const data = rowData as ArticleTableData
+
+            const confirmed = window.confirm(`Delete article '${data.title}'?`)
+            if (!confirmed) return
+
             try {
-              await api.delete(data.id);
+              await api.delete(data.id)
               setTableData((tableData) =>
                 tableData.filter((article) => article.id !== data.id)
-              );
+              )
             } catch (err) {
-              console.log(err);
+              console.log(err)
             }
           },
+        },
+        {
+          position: "toolbar",
+          icon: "add",
+          tooltip: "New article",
+          onClick: () => router.push("/tina/new"),
         },
       ]}
       options={{
         draggable: false,
         actionsColumnIndex: -1,
         filtering: true,
+        pageSize: 10,
       }}
     />
-  );
-};
+  )
+}
